Update the package to version 0.0.12 and enhance the MingdaoYun SDK
This commit addresses three primary tasks. The version of the SDK has been updated from 0.0.11 to 0.0.12 in the setup file. In the MingdaoYun.py file, a COUNT_URL was added, along with implicit renaming and usage changes of the "rowId" parameter for the delete function. In the README.md file, usage examples and an API usage guide were included to enhance clarity and aid user functionality. The changes enhance the existing MingdaoYun SDK with a new counting ability, revised deletion ability, clearer instructions and an update to the latest version.